+++ tests/testthat/test-labelled_spss.R
@@ -1,7 +1,7 @@
 context("labelled_spss")
 
 test_that("constructor checks na_value", {
-  expect_error(labelled_spss(1:10, na_values = "a"), "must be same type")
+  expect_incompatible_type(labelled_spss(1:10, na_values = "a"))
 })
 
 test_that("constructor checks na_range", {
@@ -16,7 +16,7 @@ test_that("constructor checks na_range", {
 test_that("printed output is stable", {
   x <- labelled_spss(
     1:5, c("Good" = 1, "Bad" = 5),
-    na_value = c(1, 2),
+    na_values = c(1, 2),
     na_range = c(3, Inf)
   )
   expect_output_file(print(x), "labelled-spss-output.txt")
@@ -25,7 +25,7 @@ test_that("printed output is stable", {
 test_that("subsetting preserves attributes", {
   x <- labelled_spss(
     1:5, c("Good" = 1, "Bad" = 5),
-    na_value = c(1, 2),
+    na_values = c(1, 2),
     na_range = c(3, Inf),
     label = "Rating"
   )
@@ -48,4 +48,12 @@ test_that("values in na_range flagged as missing", {
 test_that("values in na_values flagged as missing", {
   x <- labelled_spss(1:5, c("a" = 1), na_values = c(1, 3, 5))
   expect_equal(is.na(x), c(TRUE, FALSE, TRUE, FALSE, TRUE))
+})
+
+
+# Types -------------------------------------------------------------------
+
+test_that("combining preserves class", {
+  skip("todo")
+  expect_s3_class(vec_c(labelled_spss(), labelled_spss()), "haven_labelled_spss")
 })